+++ src/App.jsx
@@ -11,6 +11,7 @@ import EnergySelectReview from './views/EnergySelectReview'
 import SoftkeyReview from './views/SoftkeyReview'
 import TypographyReview from './views/TypographyReview'
 import DisplayFrameworkReview from './views/DisplayFrameworkReview'
+import DisplayWidgetsReview from './views/DisplayWidgetsReview'
 
 export default function App() {
   const [params] = useSearchParams()
@@ -32,6 +33,7 @@ export default function App() {
         <Route path="/softkey-review" element={<SoftkeyReview />} />
         <Route path="/typography-review" element={<TypographyReview />} />
         <Route path="/display-framework-review" element={<DisplayFrameworkReview />} />
+        <Route path="/display-widgets-review" element={<DisplayWidgetsReview />} />
         <Route path="*" element={<Navigate to="/" replace />} />
       </Routes>
     </SimulatorProvider>